requirements.txt
Twisted==14.0.2
pycrypto==2.6.1
pyOpenSSL==0.14
